Changed type of number

+++ homeework_2.py
@@ -1,5 +1,5 @@
-number_1 = int(input("Enter first number: "))
-number_2 = int(input("Enter second number: "))
+number_1 = float(input("Enter first number: "))
+number_2 = float(input("Enter second number: "))
 symbol = input("Enter symbol: ")
